Add files via upload
aritmetica-1.go is a very simple calculator model. Once a string is entered, its operators and numbers are recognized and the arithmetic operation is then attempted. In this version, everything is contained in a single main menu.
aritmetica-2.go incorporates the concept of functions, which are called to obtain a list of the entered numbers and their arithmetic operators, then attempt the calculation and obtain the result.